Revert "Merge branch 'kokkos:develop' into persistence"
This reverts commit 777087adabdc1aebad47b39c201aee9f7f7f6d10, reversing
changes made to 76244e04a1f22100b1ed25228da53a61dd81133a.

Merge broke it, reverting for now
