+++ perf_tests/mpi/test_main.cpp
@@ -14,9 +14,10 @@
 //
 //@HEADER
 
-#include <mpi.h>
-#include <benchmark/benchmark.h>
+#include "KokkosComm/mpi/impl/include_mpi.hpp"
+
 #include <Kokkos_Core.hpp>
+#include <benchmark/benchmark.h>
 
 // This reporter does nothing.
 // We can use it to disable output from all but the root process
