+++ src/KokkosComm/mpi/allgather.hpp
@@ -16,12 +16,11 @@
 
 #pragma once
 
-#include <mpi.h>
 #include <Kokkos_Core.hpp>
 
-#include <KokkosComm/concepts.hpp>
-#include <KokkosComm/traits.hpp>
-
+#include "KokkosComm/traits.hpp"
+#include "impl/pack_traits.hpp"
+#include "impl/include_mpi.hpp"
 #include "impl/types.hpp"
 
 namespace KokkosComm::mpi {
@@ -71,6 +70,8 @@ void allgather(const RecvView &rv, MPI_Comm comm) {
 template <KokkosExecutionSpace ExecSpace, KokkosView SendView, KokkosView RecvView>
 void allgather(const ExecSpace &space, const SendView &sv, const RecvView &rv, MPI_Comm comm) {
   Kokkos::Tools::pushRegion("KokkosComm::Mpi::allgather");
+  using SPT = KokkosComm::PackTraits<SendView>;
+  using RPT = KokkosComm::PackTraits<RecvView>;
 
   if (!KokkosComm::is_contiguous(sv) || !KokkosComm::is_contiguous(rv)) {
     throw std::runtime_error("allgather for non-contiguous views not implemented");
