+++ unit_tests/mpi/test_isendrecv.cpp
@@ -14,11 +14,10 @@
 //
 //@HEADER
 
+#include <gtest/gtest.h>
 #include <type_traits>
 
-#include <gtest/gtest.h>
-
-#include <KokkosComm/KokkosComm.hpp>
+#include "KokkosComm/KokkosComm.hpp"
 
 namespace {
